requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
loguru
slowapi
pyjwt
pyotp
qrcode[pil]
python-jose[cryptography]